# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a documentação da API para o desafio.",
        "title": "Financial Exchange Challenge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/balances": {
            "post": {
                "description": "Cria um novo balance para uma conta e asset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balances"
                ],
                "summary": "Cria um novo balance",
                "parameters": [
                    {
                        "description": "Balance",
                        "name": "balance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.CreateBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BalanceListDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "record already exists or causes a conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/balances/account/{account_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balances"
                ],
                "summary": "Lista todos os balances de uma conta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BalanceListDTO"
                            }
                        }
                    }
                }
            }
        },
        "/v1/balances/account/{account_id}/asset/{asset}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balances"
                ],
                "summary": "Busca um balance por account e asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset",
                        "name": "asset",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BalanceListDTO"
                        }
                    },
                    "404": {
                        "description": "account_id and asset are required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/balances/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balances"
                ],
                "summary": "Busca um balance por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Balance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BalanceListDTO"
                        }
                    },
                    "404": {
                        "description": "balance ID cannot be empty",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balances"
                ],
                "summary": "Atualiza um balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Balance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Balance",
                        "name": "balance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.UpdateBalanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BalanceListDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "record not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balances"
                ],
                "summary": "Deleta um balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Balance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "record not found",
                        "schema": {
                            "type": "message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "big.Float": {
            "type": "object"
        },
...
        "github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BalanceListDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "$ref": "#/definitions/big.Float"
                },
                "asset": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BigFloat": {
            "type": "object"
        },
        "github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.CreateBalanceRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "$ref": "#/definitions/github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.BigFloat"
                },
                "asset": {
                    "type": "string"
                }
            }
        },
        "github_com_mthpedrosa_financial-exchange-challenge_internal_balance_domain_dto.UpdateBalanceRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "$ref": "#/definitions/big.Float"
                }
            }
        },
...
    }
}